Add manual bid support (#17)
* Change ad set `is_autobid` parameter default to `nil` (was `true`)

This is a breaking change for current users of this library, but I think it's essential for moving forward. Rationale:

- The previous behavior (defaulting to true) is not aligned with how
  the official FB SDK works, which is not opinionated about default
  parameter values. We shouldn't be opinionated this way either.

- In the future we want to support the `bid_amount` parameter. However
  with `is_autobid` defaulting to true, if someone were to pass
  `bid_amount` without explicitly turning off autobid, the request
  would fail. That seems like an obvious developer trap that we can
  avoid by making users be explicit about autobid when they want it.

Normally a breaking change like this would require a major version
number increment, but since we're on major version 0 the public API is
not expected to be stable, so I plan to just bump the minor version
number.

* Add ability to pass `bid_amount` parameter when creating ad sets

* Version bump and changelog entry for 0.4

* Include is_autobid in list of attributes for Ad Sets

Need to include this column if we want to update its value. This should
have been part of the changes for 0.4.0 but since that isn't released
yet we can sneak this in.

* VCR cassette update for adding is_autobid